Removes GetSubcategories API Removes dead code from PythonInteractiveEvaluator Ensures only valid evaluators are persisted.
