+++ Python/Product/PythonTools/PythonTools/Options/IPythonToolsOptionsService.cs
@@ -14,14 +14,10 @@
 // See the Apache Version 2.0 License for specific language governing
 // permissions and limitations under the License.
 
-
-using System.Collections.Generic;
-
 namespace Microsoft.PythonTools.Options {
     interface IPythonToolsOptionsService {
         void SaveString(string name, string category, string value);
         string LoadString(string name, string category);
-        IList<string> GetSubcategories(string category);
         void DeleteCategory(string category);
     }
 }
