+++ Python/Product/PythonTools/PythonTools/Repl/InteractiveWindowProvider.cs
@@ -61,84 +61,6 @@ namespace Microsoft.PythonTools.Repl {
             _pythonContentType = contentTypeService.GetContentType(PythonCoreConstants.ContentType);
         }
 
-        private static string GetCategory(string evaluatorId) {
-            return SavedWindowsCategoryBase + Convert.ToBase64String(Encoding.UTF8.GetBytes(evaluatorId));
-        }
-
-        private static bool ShouldSave(IInterpreterRegistryService registry, IInteractiveEvaluator evaluator) {
-            var pyEval = evaluator as PythonInteractiveEvaluator;
-            if (pyEval == null || pyEval.Configuration == null) {
-                // Not our evaluator - don't serialize it
-                return false;
-            }
-            if (!string.IsNullOrEmpty(pyEval.ProjectMoniker)) {
-                // Directly related to a project - don't serialize it
-                return false;
-            }
-            var id = pyEval.Configuration.Interpreter.Id;
-            if (string.IsNullOrEmpty(id)) {
-                // Invalid as it has no id - don't serialize it
-                return false;
-            }
-
-            // Only serialize it if the interpreter promises to be available
-            // next time.
-            var obj = registry.GetProperty(id, "PersistInteractive");
-            return obj is bool && (bool)obj || (obj as string).IsTrue();
-        }
-
-        public void OpenPreviousWindows() {
-            var service = _serviceProvider.GetPythonToolsService();
-            var registry = _serviceProvider.GetComponentModel().GetService<IInterpreterRegistryService>();
-
-            foreach (var category in service.GetSubcategories(SavedWindowsCategoryBase)) {
-                var eval = service.LoadString("Evaluator", category);
-
-                var wnd = OpenOrCreate(eval);
-                var frame = (wnd as ToolWindowPane)?.Frame as IVsWindowFrame;
-                if (frame == null) {
-                    continue;
-                }
-            }
-        }
-
-        public void SaveAndCloseCurrentWindows() {
-            foreach (var wnd in AllOpenWindows) {
-                var frame = (wnd as ToolWindowPane)?.Frame as IVsWindowFrame;
-                if (frame == null) {
-                    continue;
-                }
-                var service = _serviceProvider.GetPythonToolsService();
-                var registry = _serviceProvider.GetComponentModel().GetService<IInterpreterRegistryService>();
-                var eval = wnd.InteractiveWindow?.Evaluator as SelectableReplEvaluator;
-                if (eval == null) {
-                    continue;
-                }
-
-                if (ShouldSave(registry, eval.Evaluator)) {
-                    //VSSETFRAMEPOS[] pos = new VSSETFRAMEPOS[1];
-                    //Guid rel;
-                    //int x, y, cx, cy;
-                    //if (ErrorHandler.Failed(frame.GetFramePos(pos, out rel, out x, out y, out cx, out cy))) {
-                    //    continue;
-                    //}
-
-                    var category = GetCategory(eval.CurrentEvaluator);
-                    service.SaveString("Evaluator", category, eval.CurrentEvaluator);
-                    //var posMode = (pos[0] & VSSETFRAMEPOS.SFP_maskFrameMode);
-                    //if (posMode == VSSETFRAMEPOS.SFP_fFloat) {
-                    //    service.SaveInt("X", category, x);
-                    //    service.SaveInt("Y", category, y);
-                    //    service.SaveInt("Width", category, cx);
-                    //    service.SaveInt("Height", category, cy);
-                    //} else if (posMode == VSSETFRAMEPOS.SFP_fTab) {
-                    //    service.SaveString("Tab", category, rel.ToString("B"));
-                    //}
-                }
-                frame.CloseFrame((uint)__FRAMECLOSE.FRAMECLOSE_NoSave);
-            }
-        }
-
         public IEnumerable<IVsInteractiveWindow> AllOpenWindows {
             get {
                 lock (_windows) {
