+++ Python/Product/PythonTools/PythonToolsService.cs
@@ -588,10 +588,6 @@ namespace Microsoft.PythonTools {
 
         #region Registry Persistance
 
-        internal IList<string> GetSubcategories(string category) {
-            return _optionsService.GetSubcategories(category);
-        }
-
         internal void DeleteCategory(string category) {
             _optionsService.DeleteCategory(category);
         }
